uploadproject/requirements.txt
Django==6.0.3
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
python-magic==0.4.27
pyclamd==0.4.0
